+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Tidarr notable changes.
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) format.
 
 ## 📦 0.4.5
+### 🚀 Added
+* [Front] A button to download all videos of an artist was added on artist page
 ### 🖍 Changed
 * [API] Run beets only for albums
 * [API] Fix undefined label in notification